Update .travis.yml
Explicitly check out the version that Travis CI is trying to build to avoid race conditions. Travis CI may think it is building one commit, but the git clone within the Docker container may end up checking out a later version if a commit occurred in the time between the build starting and the docker run command executing.